ATS_Evaluator/requirements.txt
google-generativeai
python-dotenv
pdf2image
fastapi